refactor: organize search screen components
