+++ src/components/Search/index.ts
@@ -1,0 +1,2 @@
+export { default as SearchBar } from './SearchBar'
+export { default as SearchResultList } from './SearchResultList'
